Login rework (does not work)

+++ Aqua/Kernel.cs
@@ -9,6 +9,7 @@ using term = Aqua.Terminal.Terminal;
 using Cosmos.System.Audio;
 using Cosmos.System.Audio.IO;
 using Cosmos.HAL.Drivers.PCI.Audio;
+using Aqua.Terminal.Accounts;
 
 namespace Aqua 
 {
@@ -21,18 +22,20 @@ namespace Aqua
 
         public static String currentDirectory = @"0:\";
 
-        public static AudioMixer mixer = new AudioMixer();
-        public static AC97 driver = AC97.Initialize(bufferSize: 4096);
+        public static AudioMixer mixer;
+        public static AC97 driver;
 
-        public AudioManager audioManager = new AudioManager()
-        {
-            Stream = mixer,
-            Output = driver
-        };
+        public AudioManager audioManager;
+
+        public static AccountManager accountManager;
+        public static Account curAccount;
 
         public void FirstRun()
         {
-            Console.WriteLine("Setting up things for you...\n");
+            term.DebugWrite("Setting up things for you...\n", 6);
+
+            /*if (accountManager.accounts.Count == 0)
+                curAccount = new Account("Guest", Roles.Guest);*/
 
             if (!System.IO.Directory.Exists(@"0:\Setup"))
             {
@@ -55,10 +58,13 @@ namespace Aqua
 
             try
             {
-                System.IO.File.Create(@"0:\Setup\FirstRun.acf");
-                System.IO.File.WriteAllText("0:\\Setup\\FirstRun.acf", "true");
-
-                term.DebugWrite("Successfully made the FirstRun.acf file.", 1);
+                if (!System.IO.File.Exists(@"0:\Setup\FirstRun.acf"))
+                {
+                    System.IO.File.Create(@"0:\Setup\FirstRun.acf");
+                    System.IO.File.WriteAllText("0:\\Setup\\FirstRun.acf", "true");
+                    
+                    term.DebugWrite("Successfully made the FirstRun.acf file.", 1);
+                }
             }
             catch (Exception ex)
             {
@@ -74,7 +80,23 @@ namespace Aqua
             try
             {
                 VFSManager.RegisterVFS(fs);
-                audioManager.Enable();
+                accountManager = new AccountManager();
+
+                curAccount = new Account("Guest", Roles.Guest);
+
+                if (!Sys.VMTools.IsVMWare)
+                {
+                    driver = AC97.Initialize(bufferSize: 4096);
+                    mixer = new AudioMixer();
+
+                    audioManager = new AudioManager()
+                    {
+                        Stream = mixer,
+                        Output = driver
+                    };
+
+                    audioManager.Enable();
+                }
             }
             catch (Exception ex)
             {
@@ -88,17 +110,26 @@ namespace Aqua
             Cosmos.HAL.Global.PIT.Wait(250);
             Console.Clear();
 
-            if (System.IO.File.Exists("0:\\Setup\\FirstRun.acf") && System.IO.File.ReadAllText("0:\\Setup\\FirstRun.acf") == "true")
+            if (System.IO.File.Exists("0:\\Setup\\FirstRun.acf") && System.IO.File.ReadAllText("0:\\Setup\\FirstRun.acf") == "true" || accountManager.accounts.Count == 0)
                 FirstRun();
 
             Cosmos.HAL.Global.PIT.Wait(750);
-            Aqua.Terminal.Login.LoginSystem.Start();
+            
+            /*if (accountManager.accounts.Count != 0)
+                // TO DO : https://github.com/DogOSdev/DogOS/blob/d4e3e1aeca319585f37068f3d33baa3ed94def16/DogOS/Kernel.cs
+                Aqua.Terminal.Login.LoginSystem.Start();*/
         }
 
         protected override void Run()
         {
             Console.ForegroundColor = ConsoleColor.DarkGreen;
-            Console.Write(Terminal.Login.LoginSystem.username + " | " + currentDirectory);
+            Console.Write(Terminal.Login.LoginSystem.username);
+
+            Console.ForegroundColor = ConsoleColor.White;
+            Console.Write(" @ ");
+
+            Console.ForegroundColor = ConsoleColor.Yellow;
+            Console.Write(currentDirectory);
 
             Console.ForegroundColor = ConsoleColor.Green;
             Console.Write(" $=> ");
@@ -117,8 +148,12 @@ namespace Aqua
         {
             try
             {
+                term.DebugWrite("Setting up the network...", 1);
+
                 // Setup the network / Generate an IP address dynamically
                 Network.Network.Setup();
+
+                Console.Clear();
             }
             catch (Exception e)
             {
@@ -152,7 +187,8 @@ namespace Aqua
             Console.ForegroundColor = ConsoleColor.Green;
             Console.WriteLine(devStatus[1]);
 
-            Sounds.Sounds.StartupSound();
+            if (!Sys.VMTools.IsVMWare)
+                Sounds.Sounds.StartupSound();
         }
     }
 }
